clean without breaking version control, thanks Jose

+++ gluon/admin.py
@@ -104,7 +104,7 @@ def app_cleanup(app, request):
     if os.path.exists(path):
         for f in os.listdir(path):
             try:
-                os.unlink(os.path.join(path,f))
+                if f[:1]!='.': os.unlink(os.path.join(path,f))
             except IOError:
                 r = False
 
@@ -113,7 +113,7 @@ def app_cleanup(app, request):
     if os.path.exists(path):
         for f in os.listdir(path):
             try:
-                recursive_unlink(os.path.join(path,f))
+                if f[:1]!='.': recursive_unlink(os.path.join(path,f))
             except IOError:
                 r = False
 
@@ -122,7 +122,7 @@ def app_cleanup(app, request):
     if os.path.exists(path):
         for f in os.listdir(path):
             try:
-                os.unlink(os.path.join(path,f))
+                if f[:1]!='.': os.unlink(os.path.join(path,f))
             except IOError:
                 r = False
     return r
